refactor(composables): updated getters
