+++ packages/composables/src/getters/categoryGetters.ts
@@ -1,47 +1,13 @@
 import { CategoryGetters, AgnosticCategoryTree, AgnosticBreadcrumb } from '@vue-storefront/core';
 import { Category } from '@vue-storefront/magento-api';
-import { htmlDecode } from '../helpers/htmlDecoder';
-
-const buildTree = (rootCategory: Category, currentCategory: string, withProducts = false): AgnosticCategoryTree => {
-  const hasChildren = Array.isArray(rootCategory.children) && rootCategory.children.length > 0;
-  const isCurrent = rootCategory.uid === currentCategory;
-  const label = htmlDecode(rootCategory.name);
-  const slug = `/${rootCategory.url_path}${rootCategory.url_suffix || ''}`;
-  const childrenUid = hasChildren
-    ? rootCategory
-      .children
-      .reduce((acc, curr) => [...acc, curr.uid], [])
-    : [];
-
-  const childProductCount = hasChildren
-    ? rootCategory
-      .children
-      .reduce((acc, curr) => acc + curr.product_count, 0)
-    : 0;
-
-  const items = hasChildren
-    ? rootCategory
-      .children
-      .filter((c) => (withProducts ? c.product_count > 0 : true))
-      .map((c) => buildTree(c, currentCategory))
-    : [];
-
-  return {
-    label,
-    slug,
-    uid: [rootCategory.uid, ...childrenUid],
-    items: items.filter((c) => c.count > 0),
-    count: childProductCount || rootCategory.product_count,
-    isCurrent,
-  };
-};
+import { buildCategoryTree } from '../helpers/buildCategoryTree';
 
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
 export const getTree = (category: Category): AgnosticCategoryTree | null => {
   if (!category) {
     return null;
   }
-  return buildTree(category, '');
+  return buildCategoryTree(category, '');
 };
 
 export const getCategoryTree = (
@@ -50,7 +16,7 @@ export const getCategoryTree = (
   withProducts = false,
 ): AgnosticCategoryTree | null => (
   category
-    ? buildTree(category, currentCategory, withProducts)
+    ? buildCategoryTree(category, currentCategory, withProducts)
     : null
 );
