+++ packages/composables/src/getters/userGetters.ts
@@ -1,19 +1,19 @@
 /* istanbul ignore file */
 import { UserGetters as BaseUserGetters } from '@vue-storefront/core';
-import { User } from '../types';
+import { Customer } from '@vue-storefront/magento-api';
 
-export const getFirstName = (user: User): string => (user ? user.firstname : '');
+export const getFirstName = (user: Customer): string => (user ? user.firstname : '');
 
-export const getLastName = (user: User): string => (user ? user.lastname : '');
+export const getLastName = (user: Customer): string => (user ? user.lastname : '');
 
-export const getEmailAddress = (user: User): string => (user ? user.email : '');
+export const getEmailAddress = (user: Customer): string => (user ? user.email : '');
 
-export const getRemoteShoppingAssistance = (user: User): boolean => (user ? user.allow_remote_shopping_assistance : false);
+export const getRemoteShoppingAssistance = (user: Customer): boolean => (user ? user.allow_remote_shopping_assistance : false);
 
-export const getFullName = (user: User): string => (user ? `${user.firstname} ${user.lastname}` : '');
+export const getFullName = (user: Customer): string => (user ? `${user.firstname} ${user.lastname}` : '');
 
-export interface UserGetters extends BaseUserGetters<User>{
-  getRemoteShoppingAssistance(user: User): boolean;
+export interface UserGetters extends BaseUserGetters<Customer>{
+  getRemoteShoppingAssistance(user: Customer): boolean;
 }
 
 const userGetters: UserGetters = {
